# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  internal_handlers.AddUserAccessRequest:
    properties:
      username:
        type: string
    required:
    - username
    type: object
...
  internal_handlers.CreateAccessRequest:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@omninudge.com
    name: OmniNudge Support
  description: OmniNudge social platform API
  license:
    name: Proprietary
  termsOfService: http://omninudge.com/terms/
  title: OmniNudge API
  version: "1.0"
paths:
  /access-requests/{request_id}/approve:
    post:
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Approve access request
      tags:
      - Moderation
  /access-requests/{request_id}/deny:
    post:
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Deny access request
      tags:
      - Moderation
...
  /hubs/{name}/access-request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      - description: Optional message
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handlers.CreateAccessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create hub access request
      tags:
      - Hubs
  /hubs/{name}/access-request/status:
    get:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Check access request status
      tags:
      - Hubs
  /hubs/{name}/active-users:
    get:
      parameters:
      - description: Hub name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get hub active users
      tags:
      - Hubs
...
  /mod/hubs/{hub_name}/access-requests:
    get:
      parameters:
      - description: Hub name
        in: path
        name: hub_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get pending access requests
      tags:
      - Moderation
  /mod/hubs/{hub_name}/access-requests/add-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hub name
        in: path
        name: hub_name
        required: true
        type: string
      - description: Username
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.AddUserAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Add user access by username
      tags:
      - Moderation
...
  /users/me/access-requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get my access requests
      tags:
      - Hubs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"